+++ localflavor/br/models.py
@@ -1,6 +1,9 @@
+# -*- coding: utf-8 -*-
+
 from django.db.models.fields import CharField
 from django.utils.translation import ugettext_lazy as _
 
+from . import validators
 from .br_states import STATE_CHOICES
 
 
@@ -18,3 +21,53 @@ class BRStateField(CharField):
         name, path, args, kwargs = super(BRStateField, self).deconstruct()
         del kwargs['choices']
         return name, path, args, kwargs
+
+
+class BRCPFField(CharField):
+    """
+    A model field for the brazilian document named of CPF (Cadastro de Pessoa Física)
+
+    .. versionadded:: 2.2
+    """
+
+    description = _("CPF Document")
+
+    default_error_messages = {
+        'invalid': _("Invalid CPF number."),
+        'max_digits': _("This field requires at most 11 digits or 14 characters."),
+    }
+
+    def __init__(self, *args, **kwargs):
+        kwargs['max_length'] = 14
+        super(BRCPFField, self).__init__(*args, **kwargs)
+        self.validators.append(validators.BRCPFValidator())
+
+
+class BRCNPJField(CharField):
+    """
+    A model field for the brazilian document named of CNPJ (Cadastro Nacional de Pessoa Jurídica)
+
+    .. versionadded:: 2.2
+    """
+
+    description = _("CNPJ Document")
+
+    def __init__(self, *args, **kwargs):
+        kwargs['max_length'] = 18
+        super(BRCNPJField, self).__init__(*args, **kwargs)
+        self.validators.append(validators.BRCNPJValidator())
+
+
+class BRPostalCodeField(CharField):
+    """
+    A model field for the brazilian zip code
+
+    .. versionadded:: 2.2
+    """
+
+    description = _("Postal Code")
+
+    def __init__(self, *args, **kwargs):
+        kwargs['max_length'] = 9
+        super(BRPostalCodeField, self).__init__(*args, **kwargs)
+        self.validators.append(validators.BRPostalCodeValidator())