+++ tests/settings.py
@@ -9,6 +9,7 @@ DATABASES = {
 
 INSTALLED_APPS = [
     'localflavor',
+    'tests.test_br',
     'tests.test_au',
     'tests.test_ec',
     'tests.test_md',